+++ cpp/subprojects/common/include/mlrl/common/rule_refinement/prediction_complete.hpp
@@ -106,7 +106,7 @@ class CompletePrediction final : public VectorDecorator<AllocatedVector<ScoreTyp
                                                                   const BitWeightVector& weights) const override;
 
         std::unique_ptr<IStatisticsSubset> createStatisticsSubset(
-          const IStatistics& statistics, const DenseWeightVector<uint32>& weights) const override;
+          const IStatistics& statistics, const DenseWeightVector<uint16>& weights) const override;
 
         std::unique_ptr<IStatisticsSubset> createStatisticsSubset(
           const IStatistics& statistics, const DenseWeightVector<float32>& weights) const override;
@@ -119,7 +119,7 @@ class CompletePrediction final : public VectorDecorator<AllocatedVector<ScoreTyp
 
         std::unique_ptr<IStatisticsSubset> createStatisticsSubset(
           const IStatistics& statistics,
-          const OutOfSampleWeightVector<DenseWeightVector<uint32>>& weights) const override;
+          const OutOfSampleWeightVector<DenseWeightVector<uint16>>& weights) const override;
 
         std::unique_ptr<IStatisticsSubset> createStatisticsSubset(
           const IStatistics& statistics,
