+++ cpp/subprojects/seco/src/mlrl/seco/statistics/statistics_provider_decomposable_dense.cpp
@@ -98,10 +98,10 @@ namespace seco {
              * @see `IStatistics::createSubset`
              */
             std::unique_ptr<IStatisticsSubset> createSubset(const CompleteIndexVector& outputIndices,
-                                                            const DenseWeightVector<uint32>& weights) const override {
+                                                            const DenseWeightVector<uint16>& weights) const override {
                 std::unique_ptr<DenseConfusionMatrixVector<uint32>> totalSumVectorPtr =
                   std::make_unique<DenseConfusionMatrixVector<uint32>>(this->getNumOutputs(), true);
-                return std::make_unique<StatisticsSubset<DenseWeightVector<uint32>, CompleteIndexVector, uint32>>(
+                return std::make_unique<StatisticsSubset<DenseWeightVector<uint16>, CompleteIndexVector, uint32>>(
                   std::move(totalSumVectorPtr), *this->statePtr_, *this->ruleEvaluationFactory_, weights,
                   outputIndices);
             }
@@ -110,10 +110,10 @@ namespace seco {
              * @see `IStatistics::createSubset`
              */
             std::unique_ptr<IStatisticsSubset> createSubset(const PartialIndexVector& outputIndices,
-                                                            const DenseWeightVector<uint32>& weights) const override {
+                                                            const DenseWeightVector<uint16>& weights) const override {
                 std::unique_ptr<DenseConfusionMatrixVector<uint32>> totalSumVectorPtr =
                   std::make_unique<DenseConfusionMatrixVector<uint32>>(this->getNumOutputs(), true);
-                return std::make_unique<StatisticsSubset<DenseWeightVector<uint32>, PartialIndexVector, uint32>>(
+                return std::make_unique<StatisticsSubset<DenseWeightVector<uint16>, PartialIndexVector, uint32>>(
                   std::move(totalSumVectorPtr), *this->statePtr_, *this->ruleEvaluationFactory_, weights,
                   outputIndices);
             }
@@ -203,11 +203,11 @@ namespace seco {
              */
             std::unique_ptr<IStatisticsSubset> createSubset(
               const CompleteIndexVector& outputIndices,
-              const OutOfSampleWeightVector<DenseWeightVector<uint32>>& weights) const override {
+              const OutOfSampleWeightVector<DenseWeightVector<uint16>>& weights) const override {
                 std::unique_ptr<DenseConfusionMatrixVector<uint32>> totalSumVectorPtr =
                   std::make_unique<DenseConfusionMatrixVector<uint32>>(this->getNumOutputs(), true);
                 return std::make_unique<
-                  StatisticsSubset<OutOfSampleWeightVector<DenseWeightVector<uint32>>, CompleteIndexVector, uint32>>(
+                  StatisticsSubset<OutOfSampleWeightVector<DenseWeightVector<uint16>>, CompleteIndexVector, uint32>>(
                   std::move(totalSumVectorPtr), *this->statePtr_, *this->ruleEvaluationFactory_, weights,
                   outputIndices);
             }
@@ -217,11 +217,11 @@ namespace seco {
              */
             std::unique_ptr<IStatisticsSubset> createSubset(
               const PartialIndexVector& outputIndices,
-              const OutOfSampleWeightVector<DenseWeightVector<uint32>>& weights) const override {
+              const OutOfSampleWeightVector<DenseWeightVector<uint16>>& weights) const override {
                 std::unique_ptr<DenseConfusionMatrixVector<uint32>> totalSumVectorPtr =
                   std::make_unique<DenseConfusionMatrixVector<uint32>>(this->getNumOutputs(), true);
                 return std::make_unique<
-                  StatisticsSubset<OutOfSampleWeightVector<DenseWeightVector<uint32>>, PartialIndexVector, uint32>>(
+                  StatisticsSubset<OutOfSampleWeightVector<DenseWeightVector<uint16>>, PartialIndexVector, uint32>>(
                   std::move(totalSumVectorPtr), *this->statePtr_, *this->ruleEvaluationFactory_, weights,
                   outputIndices);
             }
@@ -276,8 +276,8 @@ namespace seco {
              * @see `IStatistics::createWeightedStatistics`
              */
             std::unique_ptr<IWeightedStatistics> createWeightedStatistics(
-              const DenseWeightVector<uint32>& weights) const override {
-                return std::make_unique<WeightedStatistics<DenseWeightVector<uint32>, uint32>>(
+              const DenseWeightVector<uint16>& weights) const override {
+                return std::make_unique<WeightedStatistics<DenseWeightVector<uint16>, uint32>>(
                   *this->statePtr_, *this->ruleEvaluationFactory_, weights);
             }
